Finish moving things out to classes
